fix comments and swap broken menu for buttons
